fix:dev 算法修改提交

+++ ruoyi-modules/ruoyi-hrp/src/main/java/org/dromara/hrp/domain/HrpUserAvailability.java
@@ -37,7 +37,7 @@ public class HrpUserAvailability extends BaseEntity {
     /**
      * 星期几 (1=周一, 2=周二, ..., 7=周日)
      */
-    private Long dayOfWeek;
+    private Integer dayOfWeek;
 
     /**
      * 可上班的开始时间
