+++ ruoyi-modules/ruoyi-hrp/src/main/java/org/dromara/hrp/domain/vo/HrpUserAvailabilityVo.java
@@ -49,7 +49,8 @@ public class HrpUserAvailabilityVo implements Serializable {
      * 星期几 (1=周一, 2=周二, ..., 7=周日)
      */
     @ExcelProperty(value = "星期几 (1=周一, 2=周二, ..., 7=周日)")
-    private Long dayOfWeek;
+    private Integer dayOfWeek;
+
 
     /**
      * 可上班的开始时间
